Get rid of 'paxbin' hack and add support for user.pax.flags
The 'paxbin' hack is ugly, as it treats arguments to paxctl as if they
were for chpax and vice versa. It's also an unneded shell script that
gets (re)generated on-the-fly without real need.

On top of that, the xattr variant of PaX flags is still not supported
and the PT_PAX_FLAGS program header creation fails for some binaries and
has no fallback.

Fix all these issues by writing a proper shell script that gets passed
the flags and path and attempts to set PaX flags in all currently known
and supported ways.

Provided a required binary is available, it unconditionally:
- sets the user.pax.flags extended file system attribute,
- creates a PT_PAX_FLAGS program header or, if that fails, converts a
  GNU_STACK one and
- uses chpax resp. the locally build ./chpax to set the EI_PAX header.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>