Prevent compile time buffer overflow detection
Make the compiler loose its knowledge about the destination buffer
pointer to prevent compile time buffer overflow detection.

We use a similar trick as for __use() but this time tell the compiler
that the variable will be modified by the (empty) inline asm.

This not only fixes the related compiler warnings, it also fixes a false
positive report of 'Killed' as, apparently, gcc makes use of this UB
opportunity to let the process die.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>