Handle unexpected returns of doit()
doit(), if successful, shouldn't return. All implementations call
itworked() -- either directly or indirectly. itworked(), OTOH, calls
exit(), so never returns. This, in turn, should make doit() never return
as well.

The surrounding code, currently, doesn't account for doit() to return
and implicitly assumes it doesn't. However, as noticed in commit
3358593bf95b ("Prevent compile time buffer overflow detection"), gcc can
detect buffer overflows during compile time and exploit this UB by
simply removing all related code, breaking the toy exploit chain. This,
in combination with an odd thread tear down, leads to false positives.

As noted in the man page of pthread_kill(), a signal leading to thread
termination will affect the whole process, making the caller of
pthread_kill() *not* return and *not* being able to call exit(). This
will make the parent process see the child getting killed, leading to a
false positive "Killed" message (albeit technically correct, the process
wasn't killed by the kernel, i.e. not by PaX, but "accidentally" killed
it self).

Fix this by switching to pthread_cancel() to stop the dummy thread and
emit an explicit error message that the return of doit() is unexpected.
Getting there likely means the compiler detected one of the bugs and
dropped it in the name of UB.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>