Drop bit-rotten debian/ subdirectory
Move the only sensible file from debian/, the manpage, to a dedicated
docs/ directory and slightly update it along the way. Remove the rest.

This package is meant to be build on the host to run the tests on, as
otherwise PaX flags markings may get lost (the xattr variant).

The debian/ directory was also lacking behind (changelog not updated,
'rules' file referring to a non-existing makefile, stub documentation,
etc.). Just get rid of it all instead of kicking the dead horse any
further.

The copyright is kept intact by the means of COPYING and README.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>