Add 'big library' ASLR test
Test for huge page aligned ASLR, as semi-recent versions of Linux try to
align mappings on a PMD boundary if possible (see [1,2]), reducing the
number of address bits to guess.

Use a dedicated library for this test instead of relying on libc.so to
fulfill the size and alignment constraints.

[1] https://zolutal.github.io/aslrnt/
[2] https://git.kernel.org/linus/1854bc6e2420

Signed-off-by: Mathias Krause <minipli@grsecurity.net>